Allow generated IDL files (#17)
* Allow generated IDL files

Since the custom command is in a subdirectory it cannot depend on the
generated files directly. Instead it depends on a target from the parent
directory which depends on the IDL files.

* Allow generated IDL files

Since the custom command is in a subdirectory it cannot depend on the
generated files directly. Instead it depends on a target from the parent
directory which depends on the IDL files.

* Add comment explaining target dependency