Make the message __repr__ for python look nicer.
Before this patch, publishing a message with "ros2 topic pub"
would print something like:

publishing #5: my_msgs.msg.my_msg(axes=array('f', [0.0]))

While that is OK, it is kind of ugly.  This patch hides the
typecode and the "array" so that the output looks like:

publishing #5: my_msgs.msg.my_msg(axes=[0.0])

Which is also how it looked in Crystal.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>