Revamp how we check for the correct class.
In an issue it was pointed out that we are decref'ing
before we actually use some of the data, which means that
in theory the garbage collector could reclaim the data
before we used it.  So the original point of this change
was to fix that issue.

However, while looking at it I realized we could slightly
improve performance here by avoiding a copy of the class
and module into a combined string.  Instead, we can compare
them separately, which should reduce the copies.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>